+++ src/components/super/i-block/state/index.ts
@@ -17,7 +17,7 @@ import type Async from 'core/async';
 import type { BoundFn } from 'core/async';
 
 import { initGlobalEnv } from 'core/env';
-import { component, remoteState, hook, Hook, State } from 'core/component';
+import { component, remoteState, hook, hydrationStore, Hook, State } from 'core/component';
 
 import type bRouter from 'components/base/b-router/b-router';
 import type iBlock from 'components/super/i-block/i-block';
@@ -47,7 +47,14 @@ export default abstract class iBlockState extends iBlockMods {
 	isReadyOnce: boolean = false;
 
 	/**
-	 * A link to an application state object located in `core/component/state`.
+	 * True if the component is in the context of SSR or hydration
+	 */
+	get isRelatedToSSR(): boolean {
+		return SSR || HYDRATION && hydrationStore.has(this.componentId);
+	}
+
+	/**
+	 * A link to an application state object located in `core/component/client-state`.
 	 *
 	 * This object is used to set any general application parameters. For example, the status of user authorization or
 	 * online connection; global sharable application data, etc.
@@ -453,7 +460,6 @@ export default abstract class iBlockState extends iBlockMods {
 		return initGlobalEnv(env);
 	}
 
-	@hook({beforeRuntime: {functional: false}})
 	protected override initBaseAPI(): void {
 		super.initBaseAPI();