+++ src/core/component/client-state/index.ts
@@ -1,0 +1,10 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+export * from 'core/component/state';
+export { default } from 'core/component/state';